afrilux_sav/requirements.txt
Django>=4.2,<6.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
openai>=1.0
requests>=2.31
google-auth>=2.29
reportlab>=4.0
openpyxl>=3.1
psycopg[binary]>=3.1
argon2-cffi>=23.1
redis>=5.0
gunicorn>=22.0
